Reorganizes SidebarMenu component
